+++ NEWS.md
@@ -1,6 +1,9 @@
 rticles 0.15
 ---------------------------------------------------------------------
 
+
+- Added article template for journal *Bioinformatics*. (thanks, @ShixiangWang, #297)
+
 - Update Copernicus Publications template to version 5.8 (thanks, @nuest, #274).
 
 - Fixed issue with multi-line authors on JSS template when using `\AND`, which was firstly implemented in b740b19b90cd6f7afe2cd7d66456c9efa0bb4cdf (thanks, @aldomann, #292).